re changement du à une erreur

+++ Programme/V6/PHP/src/vue/VueAffichageItem.php
@@ -2,7 +2,7 @@
 declare(strict_types=1);
 
 namespace mywishlist\vue;
-
+use mywishlist\models\Liste;
 
 class VueAffichageItem
 {
@@ -15,17 +15,10 @@ class VueAffichageItem
 
     }
 
-    private function lesItems() : string {
-        $html = '';
-        foreach($this->tab as $item){
-            $url_item   = $this->container->router->pathFor( 'aff_item', ['id' => $item['id']] ) ;
-            $html .= "<li><a href='$url_item'>{$item['nom']}</a>,{$item['descr']}, {$item['tarif']}</li>";
-        }
-        $html = "<ul>$html</ul>";
-        return $html;
-    }
+
 
     private function unItem() : string {
+        session_start();
         $i = $this->tab[0];
 
         $url_choixdeleteItem   = $this->container->router->pathFor( 'choixdeleteItem',    ['id' => $i['id']]         ) ;
@@ -36,8 +29,11 @@ class VueAffichageItem
         $html .= "<b>Description:</b> {$i['descr']}<br>";
         $html .= "<b>Tarif:</b> {$i['tarif']}<br>";
 
-        $html .= "<a href='$url_choixmodifyItem'>Modifier</a><br>";
-        $html .= "<a href='$url_choixdeleteItem'>Supprimer</a><br>";
+        $liste= Liste::where('no','=',$i['liste_id'])->first();
+        if(isset($_SESSION['login']) && $liste['user_id']==$_SESSION['login']) {
+            $html .= "<a href='$url_choixmodifyItem'>Modifier</a><br>";
+            $html .= "<a href='$url_choixdeleteItem'>Supprimer</a><br>";
+        }
 
 
         return $html;
@@ -47,10 +43,6 @@ class VueAffichageItem
     public function render( int $select ) : string {
 
         switch ($select) {
-            case 1 : { // liste des items
-                $content = $this->lesItems();
-                break;
-            }
             case 2 : { // un item
                 $content = $this->unItem();
                 break;
@@ -60,7 +52,9 @@ class VueAffichageItem
 
         $url_accueil    = $this->container->router->pathFor( 'racine'                 ) ;
         $url_listes     = $this->container->router->pathFor( 'aff_listes'             ) ;
+        $url_deconnexion   = $this->container->router->pathFor( 'deconnexion'               ) ;
 
+        if(isset($_SESSION['login'])) {
         $html = <<<FIN
 <!DOCTYPE html>
 <html>
@@ -80,6 +74,27 @@ class VueAffichageItem
   </body>
 </html>
 FIN;
+        }else{
+            $html = <<<FIN
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Exemple</title>
+    <link rel="stylesheet" href="../CSS/design.css" />
+  </head>
+  <body>
+		<h1><a href="$url_accueil">Wish List</a></h1>
+		<nav>
+			<ul>
+				<li><a href="$url_accueil">Accueil</a></li>
+				<li><a href="$url_listes">Listes</a></li>
+			</ul>
+		</nav>
+    $content
+  </body>
+</html>
+FIN;
+        }
         return $html;
     }
